+++ ckanext/dcatde/tests/test_dcatde_serialize.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf8 -*-
+
 import rdflib
 import json
 import six